增加book表收入支出字段
还需要在服务端进行部分修改（增加删除账单时）

+++ app/src/main/java/com/lyh/guanbei/ui/fragment/CategoryChartPageFragment.java
@@ -112,7 +112,7 @@ public class CategoryChartPageFragment extends BaseFragment {
         mEntries.clear();
         List<ChartCategoryAdapter.CategoryChart> list=mAdapter.getData();
         for(ChartCategoryAdapter.CategoryChart categoryChart:list){
-            mEntries.add(new PieEntry(categoryChart.getmSum(),categoryChart.getmCategory()));
+            mEntries.add(new PieEntry((float)categoryChart.getmSum(),categoryChart.getmCategory()));
         }
     }
     public void setData(){
